feat(tagsInput): metaData Feature
The Meta Data feature allows the user to specify a list of meta data options for tags-input so that the user can specify what kind of tag they are creating.

Where $scope.metaDataOptions = [{content:'Writer'},{content:'Actor'}];

Then a drop down will appear beside each tab that will allow the user to specify what kind of tag it is -
ie: an actor, or a writer.

Adds to #2.