feat(tagsInput): Add addOnBlur option
Added tagsInput option called meta-data-type-name (that can be entered as an attribute) which is used to specify the property in the tag of the currently selected Meta Type so we can highlight it in the dropdown.

For example, if tags look like:

tag={
label:"Christian Bale"
metaType:"Actor"
}

And <tags-input meta-data="metaData"></tags-input>

and if scope.metaData = [
{
content:"Actor"
},
{content:Writer}
]

Then the meta data dropdown for the tag would have Actor highlighted
enhances #2.